+++ latest.js
@@ -1,1 +1,1 @@
-postMessage({version:'1.1.15'},'*')
+postMessage({version:'1.1.16'},'*')